Adopt Ref-based menu cursor
This allows U and D to move the cursor along with the lines,
and lets J actually delete selections, rather than padding the
end.

Currently, J sets the cursor to 1 and the pagesize to 0, which
sets the menu to the top selection.  This isn't ideal, but the
most desirable behavior is quite complex, because not only might
the line the cursor points to move, but it might also be gone
entirely.  Also, doing the right thing here will take more study
of how pageoffset and pagesize behave.  Fairly likely that I'll
circle back and implement this, because I would find it personally
annoying to lose my place in a long menu due to using J.